Урок 8. Заинжектил новый репозиторий в презентер через отдельный SubComponent (UserComponent), чтобы новый repository и dao имели время жизни как у subComponent
