--- homework5/src/main/java/com/example/homework5/dagger/AppModule.kt
@@ -1,7 +1,0 @@
-package com.example.homework5.dagger
-
-import dagger.Module
-
-@Module
-class AppModule {
-}
