+++ homework5/src/main/java/com/example/homework5/dagger/modules/AppModule.kt
@@ -1,0 +1,7 @@
+package com.example.homework5.dagger.modules
+
+import com.example.homework5.dagger.components.UserComponent
+import dagger.Module
+
+@Module (subcomponents = [UserComponent::class])
+class AppModule
